Clink v1.1.17 solves the config issue.
In the file system, the _powerline_config.lua file sorts AFTER the other
powerline scripts, and thus loads after them.  That interferes with what
can be configured by the file.

Clink v1.1.17 overrides the sort order and forces loading scripts in
this order in a directory:  first files that start with '.', then files
that start with '_', then the other files.