Major changes.
Highlights:

- The `_powerline_config.lua` file is loaded _after_ the rest of the
powerline scripts, and so several of the settings could not actually be
configured.  The way addon modules are registered has been simplified
and also redesigned so that it's possible to override the priority for
an addon module.

- A new `plc_simple` mode has been added which uses spaces between
segments and uses the fill color as the text color instead.

- All public functions are now in `plc.`, e.g. `plc.addSegment()`.

- The `plc.addSegment()` function now has an optional `right` parameter
which, when true, right-justifies the segment.

- Cleaned up the `_powerline_config.lua.sample` file and updated the
documentation.

- Color settings are now in `plc_date.`, `plc_prompt.`, etc.
E.g. `plc_prompt.textColor` and `plc_prompt.fillColor`.

- The battery module uses an updated rainbow gradient of colors when it
detects extended Xterm colors are supported.